0.1.0 - Improved Ignore System and Settings
Reworked Ignore System for Performance.

Added the ability to change if a Item is ignored or not in the Settings.

Added a Searchbar to the Settings for ease of finding a Ability.

Fixed the Settings not being able to handle missing Abilities.

Fixed all Setting Changes impacting the ExtensionBase(Basically what's used if a ability doesn't have a extension for this mod already).

Fixed Settings Not Working With VCE Content when Overriding.

Added the capacity to translate Config Settings related text.

Requires: Mastery Core 0.0.3