0.1.1 Stupid Mistakes Again
0.1.1 Stupid Mistakes Again

I forgot to account for the fact that you can't modified a list/dictionary in a foreach loop when I did a change to the settings without verifying it still worked.

Don't want to admit it but I think the previous one might have been broken when it came to abilities gaining exp or at least maybe I broke it in this true 0.1.1 instead of the previous 0.1.1 and am just too tired to remember(either way I fixed it).