0.1.1 Fixed A Idiotic Mistake and Changes to Settings
0.1.1

Fixed Vanilla Abilities not gaining exp(Originally I had named this mod Ability Proficiency in my code but I changed it because Ability Mastery sound cooler, but I forgot to change the check... didn't notice because I was mainly messing with VCE and VPE).

Visual:

If you use Mastery Core 0.0.4 and above there will now be a space between level and exp aka instead of 0(0/1000) it would be 0 (0/1000).

Settings:

now shows a Abilities translated defName and defName at the same time instead of just the defName. For ease of use and debugging.

Fixed searchbar only working for defName and not also translated defName.

Simplified the searchbar checks.

Fixed miscalculated height for opened Abilities(caused by my addition of isIgnored that I forgot to account for. Plus I forgot to account for the translation of the titles for the curves).

Removed a unrequired variable.