0.1.5 - Fixed Compatibility With VEF
Vanilla Expanded Framework even just by itself was causing issues when creating a new save when it came to pawns and gave them multiple of the same ability.

To fix this issue I had to patch GetAbility, GainAbility, and RemoveAbility. I have completely stopped the original Gain and Remove from running so that might cause compatibility issues with other mods but without this it causes even more issues for mods related to this area in the first place.

Oh I also fixed a grammar mistake in my code but I am sure you didn't need to know that.